Fixed bug where only ancester threads were displayed in 'new messages'. Also, don't mark messages as new just because they were replied to.
When we retrieve a thread matching some criteria, we also retrieve its entire tree. We build the tree structure and then store each thread in a cache indexed by id. This means that when we retrieve a subthread, e.g. because that subthread is a new message, its ancestor is cached.  But these ancestor records didn't list their subthreads correctly because we were safe-gaurding against duplicate subthreads being listed, and this safegaurd was inaccurate.