Follow up to r79941
To me it doesn't seem like there is a reason that the modules should be publicized in the api.
The modules need to be loaded locally, so the dependancy is already on local loading.

Adding a new ext.liquidThreads.edit module to do the loading of the edit mode of lqt as well as dependancy loading.