Add missing dependency and remove String.prototype.trim
The function 'finishSetup' uses mediaWiki.user.options.get, so it
depends on module 'user.options'.

While I was at it, I removed the test for mediaWiki.loader, which
seems to be unnecessary (the code is already inside of a call to
mediaWiki.using).

Use jQuery's trim function instead of changing the prototype of
String.

Change-Id: Ifbcec2adbcbcca25c1186e8ba70a81228f914238