(bug 25072) Fix weirdness in collapse threads
The problem was the use of find(), which is fully recursive, instead of
children() (now using the children ">" selector with find as per Krinkle
comments), which only search for the inmediate children. Since this
function is called recursively for .lqt-thread-replies this prevents
adding nested replies twice.

Also, modified the toggleCollapse function adding
'.lqt-post-wrapper, .lqt-thread-replies' to children() so it only
hides and shows messages and thread replies, leaving other elements that
aren't displayed anyway. This is to prevent it from also hiding the title
of the entire thread, which also contains the action links: that was
increasing the weirdness on pages with several threads.

Change-Id: I520b0451d55f4eae2bbcf795c46f975b4b0af63b