Make the collapsible control inline-block
Changed the CSS rule of the .thread-collapse-control a from display:block
to display:inline-block.

display:block was needed to be able to specify a width and height, but
this causes that, when the thread is collapsed, the text "Show X more
replies" appears below the collapsible control. Changing it to be
inline-block has the same benefits of a block element, but also the text
following it (Show X more replies) will be placed next to it instead of
below, being more intuitive to have it next to the collapse control.

Change-Id: Idd48958feedb8a93f21227f263a5466d7463891e