Add and update many PHPDoc comments
This patch intentionally touches not much code. I hope it's easy to
review because of this. It should be impossible this can break anything.

I carefully checked every type and made sure both the calling code as
well as the called code are fine with the types I'm adding in this
patch.

Change-Id: I11a4832ada3ff07732125c40f48c363dfab85b24