Fix broken CI test run
The "namespaces" section in extension.json should create the constants
as well, but it looks like this is either not executed early enough,
or this only creates the constants when a page from that namespace is
being processed. This is not the case for the failing core browser
tests, it seems.

Change-Id: I726b3a8c248bf7a72343ccfb442d82722ab87f7f