Make use of the …::class feature, where possible
This is effectively a no-op change. The …::class feature returns a
string that contains the class name. The main benefit of this change
is that tools and IDEs are able to understand that these places actually
refer to the class, instead of being random strings that contain
something that might or might not be a class name.

Change-Id: Ia47626a8b21c686081e01b0553f34c6aba32bfbe