canUserReply: Fix bad call when a Thread does not have a root
As far as I can tell this bug exists "forever". It looks like it's
possible to have a "dangling" thread that lost its parent. A lot of
code in this codebase checks if $this->root() is a valid thing,
including the Thread::title() method. But here its return value is
blindly used as if it can't be null. This patch fixes this, without
really checking why it's possible to run into this code path. I assume
what happens is that a user, bot, or script tries to reply to something
by thread ID, while that thread does not have a parent any more. Maybe
a spam script trying to reply to something that got deleted a second
ago?

Bug: T251313
Change-Id: Ic0728586842d6f8f274b63ab4713f322dd1f81f3