Remove View::getBoringSignature() call broken since 2009
The method getBoringSignature() was introduced in commit
92e7763f5f790ce38dadf102e0fa298d9429c29e (2009-10-02) by Andrew Garrett.
The author described the change as follows:

"Add a preference (on by default) as to whether or not to display custom
user signatures on posts. If it is off, it will just use the default
signature. An end to the signature wars :)."

Just a few weeks later in commit cfbcd544b3e7302c289412ca371277f7e23969dd
(2009-12-18) the same author removed the code again, but forgot about
this call. The commit message back then was:

"Make signatures a per-post affair, defaulting to the user's signature,
but editable in the process of editing a post, and indeed variable on a
per-post basis."

Luckily this code path was dead anyway because the user option
"lqtcustomsignatures" defaults to true, and there was no way to change
it.

Change-Id: I3df193f865966b7b76bd2372d6855748c7ced576