Replace deprecated MWException
permalinkData() should probably throw a checked exception. However, it
can be called in a million places that are not currently expecting any
exceptions; introducing and documenting a new exception class would add
a lot of noise in the form of `@throws` annotations that the rest of the
codebase would ignore.

So throw a proper unchecked exception instead, and duplicate the check
in the caller to avoid the exception.

Bug: T328220
Change-Id: I8b0460bfc433632d01c957fc3e5dcbfcfcae863a