Remove jquery.cookie dependency
The jquery.cookie module has been deprecated as of MediaWiki 1.41 - this
was originally going to be an update to change `jquery.cookie` to
`mediawiki.cookie` as that is the supported module now, but I searched
the repo and couldn't find any instances of jquery.cookie functionality
being used.  As such, let's delete it to remove unused dependencies

I tested this by installing the extension on my local mediawiki instance
and confirming that it removed the console warning, and that basic
functionality was not impacted.  Will also be watching client error logs
as this goes out to ensure nothing is amiss

Bug: T338637
Change-Id: I294fac21d68ccabe058d33ceb8a61fde5740f638