build: Disable MediaWiki.WhiteSpace.SpaceBeforeSingleLineComment sniff
There are 11 files with 34 findings.

This extension is unmaintenanced and has many comments after code lines.
Sniffs listed in <exclude> get autofixed by libup and indicate that the
sniff should be fixed in the long term.
Disable the sniff with serverity = 0 to indicate that
this sniff does not need fixing in this extension.

Change-Id: I02a963449626004c03a2bb63e9466ac422a44a15